requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==37.0.2
Django==4.1.1
django-anymail==8.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-email-verification==0.2.2
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
